+++ vitest.config.ts
@@ -8,6 +8,7 @@ import { defineConfig } from 'vitest/config'
 export default defineConfig({
 	test: {
 		environment: 'node',
+		setupFiles: ['src/tests/setup.ts'],
 		include: ['src/**/*.spec.ts'],
 		coverage: {
 			provider: 'v8',